# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Cinema ticket booking system with payment integration",
        "title": "Cinema Booking API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/",
    "paths": {
...
        "/api/pembayarans": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get a list of all payment records",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pembayaran"
                ],
                "summary": "Get all payments (admin only)",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Pembayaran"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new payment for a ticket",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pembayaran"
                ],
                "summary": "Create a new payment",
                "parameters": [
                    {
                        "description": "Payment details",
                        "name": "payment",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.PaymentRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/pembayarans/user/{user_id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve payment history for a user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pembayaran"
                ],
                "summary": "Get all payments for a specific user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Pembayaran"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/pembayarans/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get detailed information about a specific payment",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pembayaran"
                ],
                "summary": "Get payment details by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Payment ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Pembayaran"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update payment information including status and payment proof",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pembayaran"
                ],
                "summary": "Update payment status",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Payment ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Payment update information",
                        "name": "payment",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "bukti_pembayaran": {
                                    "type": "string"
                                },
                                "status": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Permanently delete a payment record",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pembayaran"
                ],
                "summary": "Delete a payment (admin only)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Payment ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.PaymentRequest": {
            "type": "object",
            "properties": {
                "bukti_pembayaran": {
                    "description": "URL or base64 of payment proof (optional)\nexample: https://example.com/payment-proof.jpg",
                    "type": "string"
                },
                "jumlah": {
                    "description": "The amount being paid\nexample: 50000",
                    "type": "number"
                },
                "metode_pembayaran": {
                    "description": "The payment method used\nexample: transfer_bank",
                    "type": "string"
                },
                "tiket_id": {
                    "description": "The ID of the ticket being paid for\nexample: 60d21b4667d0d8992e610c86",
                    "type": "string"
                },
                "user_id": {
                    "description": "The ID of the user making the payment\nexample: 60d21b4667d0d8992e610c84",
                    "type": "string"
                }
            }
        },
        "models.Pembayaran": {
            "type": "object",
            "properties": {
                "_id": {
                    "description": "The unique identifier of the payment\nexample: 60d21b4667d0d8992e610c88",
                    "type": "string"
                },
                "bukti_pembayaran": {
                    "description": "URL or base64 of payment proof (optional)\nexample: https://example.com/payment-proof.jpg",
                    "type": "string"
                },
                "created_at": {
                    "description": "The record creation timestamp",
                    "type": "string"
                },
                "jumlah": {
                    "description": "The payment amount\nexample: 50000",
                    "type": "number"
                },
                "metode_pembayaran": {
                    "description": "The payment method used\nexample: transfer_bank",
                    "type": "string"
                },
                "status": {
                    "description": "The payment status (pending, completed, failed)\nenum: pending,completed,failed\nexample: completed",
                    "type": "string"
                },
                "tanggal_pembayaran": {
                    "description": "The date and time when the payment was made\nexample: 2025-07-22T15:04:05Z",
                    "type": "string"
                },
                "tiket_id": {
                    "description": "The ID of the ticket associated with this payment\nexample: 60d21b4667d0d8992e610c86",
                    "type": "string"
                },
                "updated_at": {
                    "description": "The record last update timestamp",
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}